backend/requirements.txt
# requirements.txt - placeholder
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
motor>=3.3.0
pydantic>=2.5.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
livekit-api>=0.6.0
python-multipart>=0.0.6
pymongo>=4.6.0